requirement.txt
flask
python-dotenv
beautifulsoup4
slack-sdk